Move watchdog timer to utils
